Linux kernel 5.8 support (#72)
* Linux kernel 5.8 support

A block device requests queue has a function 'make_request_fn' which is
used in the driver to send requests to the device underlying to a
snapshot. Suddenly, starting from the kernel 5.8, this function has
become empty (NULL) in most cases.
Used call of the similar function 'blk_mq_make_request' in case if the
'make_request_fn' is not set. This call is wrapped into the
elastio_snap_null_mrf function with the memory barrier before call
blk_mq_make_request. Otherwise driver hangs on the snapshot destroy
operation.

Resolves #44

Note: build of the driver on the Linux kernel 5.8 has been already fixed
in the #48, commit 96d63d.

* Added possibility to use real block device in tests

Now name of the device can be specified to the tests/elio-test.sh
script as an argument.
This is useful in case of some issues related to the loop device.
Actually it's a workaround for the #71

* Enabled builds and tests for Fedora 31 with the kernel 5.8

Here is a workaround for the issue #71 has been used.
Another disk has been attached to the Fedora 31 VM especially for the
tests which are failing with the loop device (used by default).

+++ tests/devicetestcase.py
@@ -6,6 +6,7 @@
 #
 
 import os
+import subprocess
 import unittest
 
 import kmod
@@ -19,14 +20,18 @@ class DeviceTestCase(unittest.TestCase):
     def setUpClass(cls):
         cls.minor = randint(0, 23)
         r =  randint(0, 999)
-        cls.backing_store = "/tmp/disk_{0:03d}.img".format(r)
         cls.mount = "/tmp/elastio-snap_{0:03d}".format(r)
 
         cls.kmod = kmod.Module("../src/elastio-snap.ko")
         cls.kmod.load(debug=1)
-
-        util.dd("/dev/zero", cls.backing_store, 256, bs="1M")
-        cls.device = util.loop_create(cls.backing_store)
+        if os.getenv('TEST_DEVICE'):
+            cls.device = os.getenv('TEST_DEVICE')
+            dev_size = int(subprocess.check_output("blockdev --getsize64 %s" % cls.device, shell=True, text=True))//1024**2
+            util.dd("/dev/zero", cls.device, dev_size, bs="1M")
+        else:
+            cls.backing_store = "/tmp/disk_{0:03d}.img".format(r)
+            util.dd("/dev/zero", cls.backing_store, 256, bs="1M")
+            cls.device = util.loop_create(cls.backing_store)
 
         util.mkfs(cls.device)
         os.makedirs(cls.mount, exist_ok=True)
@@ -35,6 +40,7 @@ class DeviceTestCase(unittest.TestCase):
     @classmethod
     def tearDownClass(cls):
         util.unmount(cls.mount)
-        util.loop_destroy(cls.device)
-        os.unlink(cls.backing_store)
+        if hasattr(cls, "backing_store"):
+            util.loop_destroy(cls.device)
+            os.unlink(cls.backing_store)
         cls.kmod.unload()
